refactor(Twitter): Added multiple language support for `Bring back twitter`
