+++ src/main/kotlin/crimera/patches/twitter/misc/bringbacktwitter/strings/StringsMap.kt
@@ -1,0 +1,8 @@
+package crimera.patches.twitter.misc.bringbacktwitter.strings
+
+object StringsMap {
+    val replacementMap = mapOf(
+        "values" to en_rGB.values,
+        "values-en-rGB" to en_rGB.values,
+        )
+}
